checkpoint
- automatically update graphics on build
